arrangements for downstream va_list

+++ python/actionswf/__init__.py
@@ -97,7 +97,7 @@ def ImageSizes():
 def ShapeArray(*values): # ShapeArray(1,2,3)
 	return (ctypes.c_int * (len(values) + 1))(*values, 0)  # tests started from this: (ctypes.c_int * 2)(*[1, 0])
 def ActionSize():
-	return ctypes.c_ulong()  # using the .value
+	return ctypes.c_size_t()  # using the .value
 def ActionBuffer():
 	return ctypes.c_char_p()
 def Ref(a):
